+++ .run/author_monolithic in AuthorMonolithic.run.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="AuthorMonolithic in AuthorMonolithic" type="JUnit" factoryName="JUnit" nameIsGenerated="true">
+  <configuration default="false" name="author_monolithic in AuthorMonolithic" type="JUnit" factoryName="JUnit" nameIsGenerated="true">
     <module name="AuthorMonolithic" />
     <extension name="coverage">
       <pattern>
@@ -7,16 +7,11 @@
         <option name="ENABLED" value="true" />
         <option name="INCLUDE" value="false" />
       </pattern>
-      <pattern>
-        <option name="PATTERN" value="com.zemoso.author_monolithic.AuthorMonolithicApplicationTests" />
-        <option name="ENABLED" value="false" />
-        <option name="INCLUDE" value="false" />
-      </pattern>
     </extension>
     <option name="MAIN_CLASS_NAME" value="" />
     <option name="METHOD_NAME" value="" />
     <option name="TEST_OBJECT" value="directory" />
-    <dir value="$PROJECT_DIR$/src/test/java/com/zemoso/AuthorMonolithic" />
+    <dir value="$PROJECT_DIR$/src/test/java/com/zemoso/author_monolithic" />
     <method v="2">
       <option name="Make" enabled="true" />
     </method>
