+++ src/main/java/com/zemoso/author_monolithic/controller/LibraryController.txt
@@ -1,8 +1,8 @@
-package com.zemoso.AuthorMonolithic.controller;
+package com.zemoso.author_monolithic.controller;
 
-import com.zemoso.AuthorMonolithic.entity.Author;
-import com.zemoso.AuthorMonolithic.entity.Book;
-import com.zemoso.AuthorMonolithic.service.LibraryService;
+import com.zemoso.author_monolithic.entity.Author;
+import com.zemoso.author_monolithic.entity.Book;
+import com.zemoso.author_monolithic.service.LibraryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
