+++ src/main/java/com/zemoso/author_monolithic/dao/AuthorRepository.java
@@ -1,6 +1,6 @@
-package com.zemoso.AuthorMonolithic.dao;
+package com.zemoso.author_monolithic.dao;
 
-import com.zemoso.AuthorMonolithic.entity.Author;
+import com.zemoso.author_monolithic.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
