+++ src/main/java/com/zemoso/author_monolithic/dao/BookRepository.java
@@ -1,6 +1,6 @@
-package com.zemoso.AuthorMonolithic.dao;
+package com.zemoso.author_monolithic.dao;
 
-import com.zemoso.AuthorMonolithic.entity.Book;
+import com.zemoso.author_monolithic.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
