+++ src/main/java/com/zemoso/author_monolithic/entity/Book.java
@@ -1,4 +1,4 @@
-package com.zemoso.AuthorMonolithic.entity;
+package com.zemoso.author_monolithic.entity;
 
 import jakarta.persistence.*;
