+++ src/main/java/com/zemoso/author_monolithic/service/LibraryService.txt
@@ -1,9 +1,9 @@
-package com.zemoso.AuthorMonolithic.service;
+package com.zemoso.author_monolithic.service;
 
-import com.zemoso.AuthorMonolithic.dao.AuthorRepository;
-import com.zemoso.AuthorMonolithic.dao.BookRepository;
-import com.zemoso.AuthorMonolithic.entity.Author;
-import com.zemoso.AuthorMonolithic.entity.Book;
+import com.zemoso.author_monolithic.dao.AuthorRepository;
+import com.zemoso.author_monolithic.dao.BookRepository;
+import com.zemoso.author_monolithic.entity.Author;
+import com.zemoso.author_monolithic.entity.Book;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
