+++ src/test/java/com/zemoso/author_monolithic/AuthorMonolithicApplicationTests.java
@@ -1,12 +1,8 @@
-package com.zemoso.AuthorMonolithic;
+package com.zemoso.author_monolithic;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
-
-import javax.sql.DataSource;
 
 @SpringBootTest
 @ActiveProfiles("test")
