+++ src/test/java/com/zemoso/author_monolithic/dao/BookRepository.java
@@ -1,7 +1,6 @@
-package com.zemoso.AuthorMonolithic.service;
+package com.zemoso.author_monolithic.dao;
 
-import com.zemoso.AuthorMonolithic.entity.Book;
-import com.zemoso.AuthorMonolithic.dao.BookRepository;
+import com.zemoso.author_monolithic.entity.Book;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -13,13 +12,13 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-class BookServiceTest {
+class BookRepositoryTest {
 
     @Mock
     private BookRepository bookRepository;
 
     @InjectMocks
-    private BookService bookService;
+    private BookRepositoryTest bookRepositoryTest;
 
     @BeforeEach
     void setUp() {
@@ -27,7 +26,7 @@ class BookServiceTest {
     }
 
     @Test
-    void testGetBooksByAuthorId() {
+    void testFindByAuthorId() {
         Long authorId = 1L;
         Book book1 = new Book("Book1", null);
         Book book2 = new Book("Book2", null);
@@ -35,7 +34,7 @@ class BookServiceTest {
 
         when(bookRepository.findByAuthorId(authorId)).thenReturn(books);
 
-        List<Book> result = bookService.getBooksByAuthorId(authorId);
+        List<Book> result = bookRepository.findByAuthorId(authorId);
 
         assertNotNull(result);
         assertEquals(2, result.size());
@@ -45,13 +44,24 @@ class BookServiceTest {
     }
 
     @Test
-    void testSaveBook() {
+    void testSave() {
         Book book = new Book("Book1", null);
-
         when(bookRepository.save(book)).thenReturn(book);
 
-        bookService.save(book);
+        Book result = bookRepository.save(book);
 
+        assertNotNull(result);
+        assertEquals("Book1", result.getTitle());
         verify(bookRepository).save(book);
     }
+
+    @Test
+    void testDeleteById() {
+        doNothing().when(bookRepository).deleteById(1L);
+
+        bookRepository.deleteById(1L);
+
+        verify(bookRepository).deleteById(1L);
+    }
 }
+
