+++ src/test/java/com/zemoso/author_monolithic/service/BookServiceTest.java
@@ -1,6 +1,7 @@
-package com.zemoso.AuthorMonolithic.dao;
+package com.zemoso.author_monolithic.service;
 
-import com.zemoso.AuthorMonolithic.entity.Book;
+import com.zemoso.author_monolithic.entity.Book;
+import com.zemoso.author_monolithic.dao.BookRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -12,13 +13,13 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-class BookRepositoryTest {
+class BookServiceTest {
 
     @Mock
     private BookRepository bookRepository;
 
     @InjectMocks
-    private BookRepositoryTest bookRepositoryTest;
+    private BookService bookService;
 
     @BeforeEach
     void setUp() {
@@ -26,7 +27,7 @@ class BookRepositoryTest {
     }
 
     @Test
-    void testFindByAuthorId() {
+    void testGetBooksByAuthorId() {
         Long authorId = 1L;
         Book book1 = new Book("Book1", null);
         Book book2 = new Book("Book2", null);
@@ -34,7 +35,7 @@ class BookRepositoryTest {
 
         when(bookRepository.findByAuthorId(authorId)).thenReturn(books);
 
-        List<Book> result = bookRepository.findByAuthorId(authorId);
+        List<Book> result = bookService.getBooksByAuthorId(authorId);
 
         assertNotNull(result);
         assertEquals(2, result.size());
@@ -44,24 +45,13 @@ class BookRepositoryTest {
     }
 
     @Test
-    void testSave() {
+    void testSaveBook() {
         Book book = new Book("Book1", null);
+
         when(bookRepository.save(book)).thenReturn(book);
 
-        Book result = bookRepository.save(book);
+        bookService.save(book);
 
-        assertNotNull(result);
-        assertEquals("Book1", result.getTitle());
         verify(bookRepository).save(book);
     }
-
-    @Test
-    void testDeleteById() {
-        doNothing().when(bookRepository).deleteById(1L);
-
-        bookRepository.deleteById(1L);
-
-        verify(bookRepository).deleteById(1L);
-    }
 }
-
